backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
numpy==1.26.3
torch==2.1.2
torchvision==0.16.2
opencv-python==4.9.0.80
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
timm==0.9.12
matplotlib==3.8.2
scipy==1.12.0
scikit-learn==1.4.0
shapely==2.0.2
